Create openflow_messages.py
The above code will do the following:-
1. Create a topology in mininet with two nodes h1 and h2 and switch s1 and pox openflow controller is used to control switches
2. A list of 25 latest version of openflow protocol(v1.5) will be send from node h1 to h2 via switch s1.
3. To include the actual information of the OpenFlow header fields in the dump flows output, we used_openflow_msgs function to include the values of each field in the match field of the struct.pack call. 
4. The  dump-flows command will display these flows along with the actual values of these header fields. The output of dump-flows will include 25 fields, one for each header field that was sent, and the values of these fields will be displayed in the output.
5. To measure the performance,  iperf tool is used to measure the TCP and UDP bandwidth.
6. To run code we made use of 'mn' command sudo mn --custom openflow_messages.py --topo mytopo --controller remote

Note:- This will start a Mininet network with the topology defined in MyTopo class in the openflow_messages.py file, and the POX OpenFlow controller will be used to control the switches. The --controller remote option specifies that the controller will run outside of Mininet.
Once the network is started, the code will send the OpenFlow messages from h1 to h2 via s1, and print the flows in h2's switch s1. The performance metrics like time taken, TCP and UDP bandwidth, and latency can be measured using the iperf tool